Change harcoded db, fix model
This is just a horrible hack in order to prove that we can connect to an
OVSDB with a VTEP schema, and insert rows (the Open_vSwitch db name is
hardcoded, so I needed to replace it with hardware_vtep). We'll fix this
properly in a patch to submit upstream, but this allows running a simple
test that writes data to the vtep db.

Contains a patch already merged upstream related to deserialization in
OvsdbType.