Make ConfigurationService agnostic of db contents
The previous implementation assumed that the OVSDB server contained a
single OVS database, and hardcoded the "Open_vSwitch" name in most
interactions with the server. This patch removes this assumption, making
monitors, transaction, and other connection code easier to reuse.

o.o.o.l.table.internal.Tables groups the tables now by db name because
it is possible to find tables with the same names and different fields
in different schemas (e.g. Manager in the OVS and VTEP schemas).

Change-Id: Ib542f9cb556cb5b4096fefb17186cad0b40afc40
Signed-off-by: Galo Navarro <galo@midokura.com>