Get around not having OSGI service for NodeFactory
Since we're not running within OSGI it looks like we hit problems on
various pieces of the plugin that try to retrieve a Node using a
NodeFactory residing in the ServiceHelper. This patch simply allows
giving the ConfigurationService a default node, if it's can't figure out
the node by itself (our case) then it'll default to the given one.

Signed-off-by: Galo Navarro <galo@midokura.com>