Fixes ucast/mcast find MAC remote bug
This patch set fixes a bug when a search for the a unicast or
multicast remote MAC entry would return an incorrect result, because
the search method would compare a set (OvsDBSet) of logical switch
UUIDs with the UUID of the MAC logical switch. This comparison
would always return false.

This incorrect behavior affected other method which depended on the
find remote MAC functionality, such as delete remote MAC.

Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>