Fixes ConnectionService resource leak
The ConnectionService was leaking file descriptors be creating
a new netty event loop group for each new connection, without saving
and destroying the group when the client disconnected.

This patch adds a single event loop group for the connection service,
which can be closed by calling the service's destroy() method.

The patch also moves initialization code for both the ConnectionService
and the InventoryService from the public init() method to the class
constructor, preventing any caller code to re-initialize, and therefore
ovewrite resource data structures that should be cleaned up.

Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>