Fix for configuration vtepBindVlan() method
The vtepBindVlan() method creates a new port-VLAN binding for a
specified logical switch. If the logical switch does not exist,
the method will create a logical switch with the specified name
and VNI.

Creating the port-VLAN binding requires the UUID of the logical
switch, as stored in the VTEP database.

Previously, the creation of the logical switch and binding was
performed in the same transaction, with the wrong value set for
the logical switch UUID, for both the binding and the flooding IP
entries. Hence, when the logical switch did not exist, the operation
failed.

This change creates the logical switch on a different transaction,
such that the status returns the correct logical switch UUID.

Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>